@@ -1,5 +1,6 @@
-import { GenreResponse } from "@/types";
 import { fetchFromApi } from "./index";
+
+import { GenreResponse } from "@/types";
 
 export async function fetchGenres(): Promise<GenreResponse[]> {
   const data = await fetchFromApi("/genre/movie/list");
